third_party/blink/web_tests/external/wpt/html/infrastructure/urls/resolving-urls/query-encoding/location.sub_encoding=x-cp1251-expected.txt

This is a testharness.js-based test.
PASS location [PutForwards]
PASS location.assign()
PASS location.replace()
PASS location.href
FAIL location.search assert_equals: expected "?%26%23255%3B" but got "?%C3%BF"
Harness: the test ran to completion.
